--- docarray/util/__init__.py
@@ -1,3 +1,0 @@
-from docarray.util.find import find, find_batched
-
-__all__ = ['find', 'find_batched']
