+++ docarray/utils/find.py
@@ -2,9 +2,10 @@ from typing import List, NamedTuple, Optional, Type, Union
 
 from typing_inspect import is_union_type
 
-from docarray import BaseDocument, DocumentArray
 from docarray.array.abstract_array import AnyDocumentArray
+from docarray.array.array import DocumentArray
 from docarray.array.array_stacked import DocumentArrayStacked
+from docarray.document import BaseDocument
 from docarray.typing import AnyTensor
 from docarray.typing.tensor.abstract_tensor import AbstractTensor
